Day 6 - Error prevention and refactorings

+++ Forms/Favorites.cs
@@ -1,4 +1,5 @@
-﻿using cat_search.Model;
+﻿using cat_search.Extensions;
+using cat_search.Model;
 using cat_search.Services;
 using System.Net;
 
@@ -15,40 +16,66 @@ namespace cat_search.Forms
 
             apiService = new ApiService();
 
-            favourites = apiService.GetFavourites();
+            try
+            {
+                favourites = apiService.GetFavourites();
 
-            var favouritesNames = favourites.Select(f => f.Name).ToArray();
+                var favouritesNames = favourites.Select(f => f.Name).ToArray();
 
-            lboxFavourites.Items.AddRange(favouritesNames);
+                lboxFavorites.Items.AddRange(favouritesNames);
+
+                if (lboxFavorites.Items.Count == 0)
+                    btnDeleteFavorite.Enabled = false;
+            }
+            catch (Exception ex)
+            {
+                ex.ShowApplicationErrorMessageBox();
+            }
         }
 
         private void btnDeleteFavourite_Click(object sender, EventArgs e)
         {
-            var favouriteName = lboxFavourites.SelectedItem as string; //qual a diferença para toString?
-
-            int favouriteId = favourites.Where(x => x.Name == favouriteName)
-                .Select(x => x.Id)
-                .FirstOrDefault();
-
-            var response = apiService.DeleteFavourite(favouriteId);
-
-            if (response.StatusCode == HttpStatusCode.OK)
+            if (lboxFavorites.SelectedItem is not string favouriteName)
             {
-                lboxFavourites.Items.Remove(favouriteName);
-
-                MessageBox.Show($"Favorito da raça {favouriteName} excluída com sucesso!",
-                    "Sucesso",
-                    MessageBoxButtons.OK,
-                    MessageBoxIcon.Information);
-            }
-            else
-            {
-                MessageBox.Show($"Erro ao excluir favorito!\n" +
-                    $"Erro: {response.StatusDescription} -  {response.Content}",
+                MessageBox.Show("You must choose a breed to remove from the favorites list!",
                     "Aviso",
                     MessageBoxButtons.OK,
                     MessageBoxIcon.Warning);
+
+                return;
             }
+
+            try
+            {
+                int favouriteId = favourites.Where(x => x.Name == favouriteName)
+                        .Select(x => x.Id)
+                        .FirstOrDefault();
+
+                var response = apiService.DeleteFavourite(favouriteId);
+
+                if (response.StatusCode == HttpStatusCode.OK)
+                    lboxFavorites.Items.Remove(favouriteName);
+
+                response.ShowMessageBox($"Breed {favouriteName} deleted from the favorites list successfully!");
+
+                if (lboxFavorites.Items.Count == 0)
+                    btnDeleteFavorite.Enabled = false;
+            }
+            catch (Exception ex)
+            {
+                ex.ShowApplicationErrorMessageBox();
+            }
+        }
+
+        private void btnReturn_Click(object sender, EventArgs e)
+        {
+            Hide();
+
+            Search searchForm = new();
+
+            searchForm.Closed += (s, args) => Close();
+
+            searchForm.Show();
         }
     }
 }
